pyproject.toml
[project]
name = "visual-hyde"
version = "0.1.0"
description = "Visual HyDE: Cross-modal chart retrieval via hypothetical chart generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core ML
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "open-clip-torch>=2.24.0",
    # Image processing
    "pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    # Vector search
    "faiss-cpu>=1.8.0",
    # Data
    "datasets>=2.18.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    # PDF processing
    "pymupdf>=1.24.0",
    # Config & utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    # Evaluation
    "scikit-learn>=1.4.0",
    # HTTP (for model APIs)
    "httpx>=0.27.0",
    "anthropic>=0.26.0",
    "openai>=1.30.0",
]

[project.scripts]
visual-hyde = "visual_hyde.cli:app"
build-corpus = "visual_hyde.scripts.build_corpus:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/visual_hyde"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "ipykernel>=6.29.0",
    "notebook>=7.1.0",
    "black>=24.3.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
